.travis.yml: fast_finish & allow nightly to fail
fast_finish makes the build report as failed when one build (which is not allowed to fail) fails, but the other jobs still continue running.

I think you possibly want to allow php nightly to fail as it can have php nightly bugs. Tests for it are still ran anyway.

I am not sure on php 7, but I think you might want to allow it to fail too.