Large commit...
Split out the messages implementation, the command parsing subsystem
and generally touch up the code in lots of places.

The validation subsystem is broken and not implemented in commands.
The command help system is removed.

All tests are broken and the bot seems a bit unstable but it works enough.
All modules will need a refresh.

Signed-off-by: Rick Kerkhof <rick.2889@gmail.com>